+++ src/main/java/com/tiny/ledger/transaction/dto/TransactionResponse.java
@@ -2,5 +2,7 @@ package com.tiny.ledger.transaction.dto;
 
 import com.tiny.ledger.transaction.domain.TransactionType;
 
-public record TransactionResponse(Long transactionId, Double amount, TransactionType type) {
+import java.math.BigDecimal;
+
+public record TransactionResponse(Long transactionId, BigDecimal amount, TransactionType type) {
 }
